# docs/api.yaml
openapi: 3.0.3
info:
  title: FollowingFeed API
  version: 1.0.0
  description: >-
    FollowingFeed 后端接口契约。除公开接口外，请通过 Bearer Token 认证。
    成功返回 2xx，客户端错误返回 4xx，服务端错误返回 5xx；
    响应体中的 code 用于细分业务结果。除健康检查和文档路由外，业务接口均受
    Redis IP 限流保护：超限返回空响应体的 HTTP 429，限流依赖异常返回空响应体的 HTTP 500。
    需要 Bearer 认证的接口在 Redis 会话检查失败时返回空响应体的 HTTP 503。
servers:
  - url: /api/v1
...
paths:
  /users/signup:
    post:
      tags: [User]
      summary: 注册用户
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignUpRequest" }
      responses:
        "200": { $ref: "#/components/responses/NoDataResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "409": { $ref: "#/components/responses/ConflictResponse" }
        "413": { $ref: "#/components/responses/PayloadTooLargeResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /users/login:
    post:
      tags: [User]
      summary: 邮箱密码登录
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200": { $ref: "#/components/responses/LoginResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/BusinessErrorResponse" }
        "413": { $ref: "#/components/responses/PayloadTooLargeResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /users/profile:
    get:
      tags: [User]
      summary: 获取当前用户资料
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/UserResultResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /users/refreshToken:
    post:
      tags: [User]
      summary: 刷新访问令牌
      description: 浏览器自动携带登录时设置的 HttpOnly 刷新令牌 Cookie。
      security: [{ refreshCookie: [] }]
      responses:
        "200": { $ref: "#/components/responses/AccessTokenResultResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "503": { $ref: "#/components/responses/BusinessErrorResponse" }
  /users/logout:
    post:
      tags: [User]
      summary: 注销当前会话
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/NoDataResultResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /pub/users/{id}/profile:
    get:
      tags: [User]
      summary: 获取作者公开资料
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200": { $ref: "#/components/responses/PublicUserResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "404": { $ref: "#/components/responses/NotFoundResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: followingfeed_refresh_token
  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: 用户 ID
      schema: { type: integer, format: int64, minimum: 1 }
      example: 1
...
  schemas:
    SignUpRequest:
      type: object
      required: [nickname, email, password, confirm_password]
      properties:
        nickname:
          type: string
          minLength: 2
          maxLength: 6
          description: 2 至 6 位中文、英文字母或数字，且不得使用系统保留名称
          example: 云端旅人
        email:
          type: string
          format: email
          maxLength: 254
          example: alice@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          description: >-
            8 至 72 个 ASCII 字符，并包含字母、数字和以下至少一个特殊字符：
            $ @ ! % * # ? &
          example: Passw0rd!
        confirm_password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          description: 必须与 password 一致
          example: Passw0rd!
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: alice@example.com }
        password: { type: string, format: password, example: Passw0rd! }
...
    ResultMeta:
      type: object
      required: [code, msg]
      properties:
        code: { type: integer, example: 0 }
        msg: { type: string, example: success }
    NoDataResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { type: object, nullable: true, example: null }
    BusinessErrorResult:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: integer
          enum: [4, 5]
          description: 4 表示参数或权限等业务错误，5 表示服务处理失败
          example: 4
        msg: { type: string, example: 参数错误 }
        data: { type: object, nullable: true, example: null }
    UserVO:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        nickname: { type: string, example: 云端旅人 }
        email: { type: string, format: email, example: alice@example.com }
        createdAt: { type: integer, format: int64, description: Unix 毫秒 }
        updatedAt: { type: integer, format: int64, description: Unix 毫秒 }
        articleLikeCount: { type: integer, format: int64, minimum: 0 }
        articleReadCount: { type: integer, format: int64, minimum: 0 }
        articleCollectCount: { type: integer, format: int64, minimum: 0 }
    PublicUserVO:
      type: object
      required:
        [id, nickname, createdAt, followingCount, followersCount, articleCount]
      properties:
        id: { type: integer, format: int64, example: 1 }
        nickname: { type: string, example: 云端旅人 }
        createdAt: { type: integer, format: int64, description: Unix 毫秒 }
        followingCount: { type: integer, format: int64, minimum: 0 }
        followersCount: { type: integer, format: int64, minimum: 0 }
        articleCount: { type: integer, format: int64, minimum: 0 }
    UserResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/UserVO" }
    PublicUserResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/PublicUserVO" }
...
  responses:
    LoginResultResponse:
      description: 登录成功时通过响应头返回访问令牌，并设置 HttpOnly 刷新令牌 Cookie
      headers:
        X-JWT-Token:
          description: 访问令牌
          schema: { type: string }
        Set-Cookie:
          description: HttpOnly 刷新令牌 Cookie
          schema: { type: string }
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/NoDataResult"
              - $ref: "#/components/schemas/BusinessErrorResult"
    AccessTokenResultResponse:
      description: 刷新成功时通过响应头返回新的访问令牌
      headers:
        X-JWT-Token:
          description: 新的访问令牌
          schema: { type: string }
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/NoDataResult"
              - $ref: "#/components/schemas/BusinessErrorResult"
    NoDataResultResponse:
      description: 操作成功时 data 为 null
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/NoDataResult"
              - $ref: "#/components/schemas/BusinessErrorResult"
    UserResultResponse:
      description: 当前用户资料或业务错误
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/UserResult" } },
        }
    PublicUserResultResponse:
      description: 作者公开资料或业务错误
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/PublicUserResult" } },
        }
...
    BusinessErrorResponse:
      description: 与 HTTP 状态码对应的业务错误
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    BadRequestResponse:
      description: 请求格式或参数错误
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    PayloadTooLargeResponse:
      description: 请求体超过 1 MiB 限制
      content:
        application/json:
          schema: { $ref: "#/components/schemas/BusinessErrorResult" }
...
    NotFoundResponse:
      description: 请求的资源不存在
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    ConflictResponse:
      description: 请求与资源当前状态冲突
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    InternalServerErrorResponse:
      description: 服务内部错误
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    UnauthorizedResponse:
      description: Token 缺失、无效或会话已失效；响应体为空